feat: 눈덩이 컴포넌트 분리
